requirements.txt
aniso8601==9.0.1
APScheduler==3.9.1
attrs==22.1.0
click==8.1.3
Flask==2.1.2
Flask-APScheduler==1.12.4
Flask-Injector==0.14.0
flask-restx==0.5.1
injector==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2022.2
tzlocal==4.2
Werkzeug==2.1.2
